fix(evm): roll back the value transfer when a transaction fails
The runner performs the call-path value transfer outside the engine, as part
of loading pre-state, which resets the journal — so the engine's revert could
not undo it. A transaction that ran out of gas still moved its value.

Prague state tests: 43.07% -> 73.06% on an 8000-case sample (80.27% on the
first 1500). All forks: 69.69%.

Two process notes. The fix was found by bucketing balance mismatches by
`wei-delta / gasPrice`: 235 cases off by exactly one wei with a zero gas delta
pointed at value, not gas, and the fixture's unchanged recipient balance
confirmed the transaction was expected to fail. Guessing at the surcharge
rules had wasted two attempts before that.

And the first version of this fix silently did nothing: the formatter had
reflowed the call site across lines, so the patch adding the two new arguments
never matched, and I had not re-run `check:types` to catch the arity error.

Also makes the refund counter signed. EIP-3529 decrements it on a slot
revisit and the running total is allowed to go negative; clamping at zero
over-refunded.

+++ src/evm/internal/c/evm.c
@@ -1853,7 +1853,7 @@ int evm_execute_create(evm_vm *vm, int init_len, int64_t gas) {
 }
 
 EXPORT("evm_refund") int64_t evm_refund(evm_vm *vm) {
-  return (int64_t)vm->st->refund;
+  return vm->st->refund;
 }
 
 // --- post-state readback ---
